# small optimization of BytesIOPayload.size